#221 refactor: 재사용 스타일을 common 폴더로 분리
