+++ src/Pages/InvitePage.tsx
@@ -2,10 +2,10 @@ import { getSearchMember } from '@/apis';
 import { Check } from '@/assets/icons';
 import { IconWrapper } from '@/components';
 import { Button, Form } from '@/components/common';
-import { checkedStyle, checkIconStyle } from '@/components/Todo';
 import { useToastHandler } from '@/hooks';
 import { useInviteMember } from '@/hooks/mutations';
 import { useSquadStore } from '@/stores';
+import { checkedStyle, checkIconStyle } from '@/styles/common';
 import { SearchMemberResponse } from '@/types';
 import { css, Theme, useTheme } from '@emotion/react';
 import { FormEvent, KeyboardEventHandler, useState } from 'react';
