+++ src/components/Todo/styles.tsx
@@ -52,30 +52,6 @@ export const contentWrapperStyle = css`
   height: 100%;
 `;
 
-export const checkIconStyle = (theme: Theme) => css`
-  min-width: 20px;
-  max-width: 20px;
-  height: 20px;
-  background-color: ${theme.colors.background.white};
-  border: 2px solid ${theme.colors.gray.gray100};
-  border-radius: 8px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const checkedStyle = (theme: Theme) => css`
-  background-color: ${theme.colors.primary};
-  border: none;
-
-  & svg {
-    color: white;
-    stroke-width: 2;
-    width: 12px;
-    height: 12px;
-  }
-`;
-
 export const getStatusStyles = (isChecked: boolean, theme: Theme) => {
   switch (isChecked) {
     case true:
